monitoring/requirements.txt
# Monitoring Microservice Dependencies

# Web Framework
fastapi==0.123.7
uvicorn==0.38.0

# WebSocket/Socket.IO
python-socketio[asyncio]==5.12.0
eventlet==0.40.3

# Database
pymongo==4.10.1

# Message Queue
pika==1.3.2

# HTTP Client
httpx==0.28.1
requests==2.31.0

# Validation
pydantic==2.12.5
email-validator==2.3.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.2.1

# Logging
pyyaml==6.0.3

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
